+++ src/applications/ask-va/tests/e2e/fixtures/flows/inbox/authenticated/2k.yml
@@ -1,5 +1,5 @@
 flow:
-  title: Dashboard inbox page without cards (auth)
+  title: Inbox page without cards (auth)
   runOnCI: true # when this is false, the flow will only run on a scheduled job
   steps:
     - action: include
