+++ src/applications/ask-va/tests/e2e/run-yaml-tests.cypress.spec.js
@@ -17,7 +17,7 @@ import intercept3rdPartyResponses from './fixtures/api-mocks-for-3rd-party';
 import STEPS from './actions';
 
 import formsTestsToRun from './fixtures/flows/forms/tests-to-run.json';
-import dashBoardTestsToRun from './fixtures/flows/dashboard/tests-to-run.json';
+import InboxTestsToRun from './fixtures/flows/inbox/tests-to-run.json';
 
 const EMPTY_FLOW_YML = `
 flow:
@@ -231,7 +231,7 @@ describe('YAML tests', () => {
           }
         }
       };
-      runTestsForFilesInPath('dashboard', dashBoardTestsToRun);
+      runTestsForFilesInPath('inbox', InboxTestsToRun);
       runTestsForFilesInPath('forms', formsTestsToRun);
     });
   });
